backend/requirements.txt
feedparser
fastapi
uvicorn
aiosqlite
python-dotenv
openai
apscheduler